Fix #37
Parse 'collection and 'implies when we retrieve the information.
This is done in a way that is as compliant with the existing format
as much as possible for valid values. However, invalid values will be
explicitly marked / quoted so that they cause no trouble.

Two significant changes for valid values:

- a 'core in 'implies will be further wrapped in a list to distinguish
  a symbol from a string.
- a 'use-pkg-name will be further wrapped in a list to distinguish
  a symbol from a string.

These two break the compliance, but originally it's not possible to
unambiguously use this information anyway.