Fix invalid source
In version 5.3.6 (in 2013), the key 'source and 'checksum were added
to the JSON file, so it's initialized with the information from the
version exception for 5.3.6, falling back to the actual 'source if
the version exception for 5.3.6 doesn't exist.
(https://github.com/racket/pkg-index/commit/5f3be74f915c821361c53bd0ca22049b45c5a6d5)

The issue is that if a package still has the version exception, but
updated 'source to be different from the version exception,
this update will not be reflected in the JSON file in the 'source field
because it prioritizes the version exception for 5.3.6 first.

This PR fixes the issue. Also note that 'source is a mandatory field, so we
don't need to have a fallback value.

(Really, we should have fixed the database if any information is
missing, rather than adding more workaround on the program that queries the
database)

Fixes #34